Extract and export the `struct` serializer
